Backport TRUNK-6572 and TRUNK-6576 to 2.8.x
- TRUNK-6572: Add @Transactional(propagation = Propagation.NOT_SUPPORTED) to OrderServiceImpl#getNewOrderNumber
- TRUNK-6576: Add dose > 0 validation to DrugOrderValidator, add error message in messages.properties, and add tests in DrugOrderValidatorTest

+++ api/src/main/java/org/openmrs/validator/ValidateUtil.java
@@ -148,7 +148,7 @@ public class ValidateUtil {
 			if (value == null || !(value instanceof String)) {
 				continue;
 			}
-			long length = Context.getAdministrationService().getMaximumPropertyLength((Class<? extends OpenmrsObject>) aClass, field);
+			int length = Context.getAdministrationService().getMaximumPropertyLength((Class<? extends OpenmrsObject>) aClass, field);
 			if (length == -1) {
 				return;
 			}
